+++ src/app/components/ruta/rutas/rutas.component.ts
@@ -3,6 +3,7 @@ import { RutaService } from "../../../services/ruta.service";
 import { Global } from "../../../services/global";
 import { Location } from "@angular/common";
 import { HttpErrorResponse } from "@angular/common/http";
+import Swal from "sweetalert2";
 
 @Component({
   selector: "app-rutas",
@@ -63,19 +64,30 @@ export class RutasComponent implements OnInit {
   }
 
   deleterutas() {
-    this._rutaService.deleteRutas().subscribe(
-      (response) => {
-        this.message = response.message;
-        this.isAlert = true;
-        this.onIsError();
-        this.getRutas();
-      },
-      (error) => {
-        this.message = error.message;
-        this.isAlert = false;
-        this.onIsError();
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Este proceso es irreversible.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Si",
+      cancelButtonText: "No",
+    }).then((result) => {
+      if (result.value) {
+        this._rutaService.deleteRutas().subscribe(
+          (response) => {
+            this.message = response.message;
+            this.isAlert = true;
+            this.onIsError();
+            this.getRutas();
+          },
+          (error) => {
+            this.message = error.message;
+            this.isAlert = false;
+            this.onIsError();
+          }
+        );
       }
-    );
+    });
   }
   onIsError() {
     this.isError = true;
@@ -88,5 +100,20 @@ export class RutasComponent implements OnInit {
 
   goBack() {
     this._location.back();
+  }
+
+  deleteConfirmation(id: string) {
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Este proceso es irreversible.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Si",
+      cancelButtonText: "No",
+    }).then((result) => {
+      if (result.value) {
+        this.delete(id);
+      }
+    });
   }
 }
