+++ src/app/components/usuario/users/users.component.ts
@@ -4,6 +4,7 @@ import { PublicationService } from "../../../services/publication.service";
 import { Global } from "../../../services/global";
 import { Location } from "@angular/common";
 import { HttpErrorResponse } from "@angular/common/http";
+import Swal from "sweetalert2";
 
 @Component({
   selector: "app-users",
@@ -66,19 +67,30 @@ export class UsersComponent implements OnInit {
   }
 
   deleteUsers() {
-    this._userService.deleteUsers(this.resID).subscribe(
-      (response) => {
-        this.message = response.message;
-        this.isAlert = true;
-        this.onIsError();
-        this.getUsers(this.resID);
-      },
-      (error) => {
-        this.message = error.message;
-        this.isAlert = false;
-        this.onIsError();
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Este proceso es irreversible.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Si",
+      cancelButtonText: "No",
+    }).then((result) => {
+      if (result.value) {
+        this._userService.deleteUsers(this.resID).subscribe(
+          (response) => {
+            this.message = response.message;
+            this.isAlert = true;
+            this.onIsError();
+            this.getUsers(this.resID);
+          },
+          (error) => {
+            this.message = error.message;
+            this.isAlert = false;
+            this.onIsError();
+          }
+        );
       }
-    );
+    });
   }
 
   onIsError() {
@@ -92,5 +104,20 @@ export class UsersComponent implements OnInit {
 
   goBack() {
     this._location.back();
+  }
+
+  deleteConfirmation(id: string) {
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "Este proceso es irreversible.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Si",
+      cancelButtonText: "No",
+    }).then((result) => {
+      if (result.value) {
+        this.delete(id);
+      }
+    });
   }
 }
